git-sync: support squash or rebase workflows (#538)
This week I decided to start squashing PRs that have a single commit.
This results in prettier history than a --no-ff merge commit for a
single commit... The end result of this single-commit squash is similar
to rebase, except that github adds the PR number to the title the
squashed commit.

To support this new workflow, I'm updating the git-sync script. We can't
automatically delete the local copy of the branch that was squashed,
because it would require a risky `git branch --delete --force`. But we
can print a message to promt the user to delete it manually.

Also, this removes a debugging print statement that I had left by
accident in the previous commit. Oops!