pallenec: Remove --dump flag (#536)
This commit removes the rarely-used --dump pallenec command flag. If you
ain't gonna need it, better delete it...  If a Pallene developer still
wants to see this info, they can add some temporary print statements to
the driver.lua.

One thing we gain by removing this code is that it gets rid of our
dependency on the "inspect" library. Another is that the we get rid of
the "list_of_compiler_passes" array, which was always out of date.