Check that the installed Lua version is the correct one. (#541)
This adds a compile-time check to see that the installed version of Lua
matches the one that is supported by the current version of Pallene. It
also adds a run-time check that raises an error in case you update your
Lua but do not recompile your Pallene module. Fixes #18.

ATTENTION: I had to add a luaL_checkcoreversion function to the
lua-internals. This means that you must update your lua-internals.