Oops. Fix a bug where --print-ir was unusable (#542)
This fixes an embarassing bug from #536. Instead of printing the IR,
it would compile the Pallene file, as if you had called pallenec with
no command line flags. Unfortunately, the unit tests didn't catch this
because we were only looking at the exit code and not whether it was
generating correct IR output.

I still don't think that the test suite should look at the entire IR,
because that changes often and would break the tests all the time.
However, looking for the "function main" that is always there at least
ensures that the output isn't completely empty, which is what was
happening here.