Move the body of the pallenec to the src/ directory
We had left the pallenec out of the list of Lua files in the run-lint,
which allowed it to accumulate a bunch of Lint errors. To stop this from
happening again, I moved the bulk of the pallenec into a module inside
the src/ directory, where the linter can see it.

I considered just telling the linter to add the pallenec to the list
but that was a bit awkward. Because the pallenec doesn't have a ".lua"
extension, it would not be matched by the globs.