Avoid gcc warning for math.mininteger literals
Running the test suite under -Wall -Werror would choke on programs with
-9223372036854775808 as an integer literal. GCC warns that the "integer
constant is so large that it is unsigned". I am not sure what version of
GCC added this warning and/or enabled it in -Wall, but it happens for me
on GCC 12.1.1.

The fix is to represent these things using the appropriate C macros,
similarly to how we handle HUGE_VAL for floating point numbers.