Fix a bug where the stack would not grow when it should
We were using the luaD_checkstack macro incorrectly. That macro asks for
the total height of the Lua stack, all the way to the Lua entry point.
The result of the bug was that we would not grow the Lua stack when we
should, which could surface as a segmentation fault.

While debugging this, I found that address sanitizer was very helpful.
I have also added some documentation describing how to run Pallene with
address sanitizer.