More flexible consistency rule
This commit loosens the type consistency restriction for type casts.
Previously we used the rule from Gradual Typing, which required that
both types be equal up to 'any'. In the new rule we only check if the
type tag matches. Casting to and from "any" is still allowed.
Furthermore, the new rule also allows casts from one array type to
another, even if the array contents are different. We also between any
pair of function type no matter the type of argument or arity.

The motivation for this rule is that Pallene type checks only check the
type tag anyway. The type casts that used to be forbidden but are now
allowed are type casts that theoretically might run without error at
runtime. Now we only forbit type casts that are guaranteed to fail, such
as int=>string or array=>function.