Replace typedecl.match_tag by typedecl.typename
The match_tag API would have been a clever way to implement "instanceof" for
a multi-level class hierarchy. However, we don't actually need to do that...

The new function maintains a hash table that maps the full tag name (ast.Exp.Num)
to the type name (ast.Exp). Compared to the old match_tag, this is faster and
needs no string allocations. We will also give an error if someone passes a
type name that does not exist.