Remove -Wpedantic flag from run-tests script
We do not need -Wpedantic as we are specifying C99 standard. According to GCC, -Wpedantic will generate warnings based on strict ISO C/ANSI C, which is basically C89. If we really need to use some features from C89 (ISO C), we may use flags specific to the feature we want to use e.g. -Wdeclaratoin-after-statement to lift all the variable declarations at the top of the function definition etc.