Stop using /**/ to signal blank lines
In-band signaling is evil :)

The original idea behind /**/ was that it was difficult to control when
the coder generated a blank line, so we told the reindenter to delete
all blank lines unless they were specially marked with /**/. However,
it turns out that it's not too difficult to avoid unwanted blank lines.
The main problem used to be situations where we did a table.concat("\n")
where some elements in the list already have a newline. The fix is the
`concat_lines` function, which removes said newlines before we concat.